requirements.txt
flask
flask-wtf
python-dotenv
duckdb